allora: a. aggiunti repository storici; b. sistemato metodo arrotondamento che legge anche se il bagno deve o meno essere approsimato e quanto o se non deve fare alimentazione. c. sistemato OperazioniBagnoService con stream, lambda e robe strane varie ma che, in teoria, dovrebbero funzionare. vedremo poi se anche nella pratica sarà cosi.

+++ src/main/java/org/galvanica/math/ScattiMath.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 
 public class ScattiMath {
-    private int restoScatti;
+    private long restoScatti;
     private double moltiplicatoreAlimentazione;
 
 }
